+++ test/fixtures/workflow_steps.exs
@@ -31,7 +31,7 @@ defmodule Test.WorkflowSteps do
   defp checkout do
     [
       name: "Checkout",
-      uses: "actions/checkout@v2"
+      uses: "actions/checkout@v3"
     ]
   end