+++ test/git_hub_actions/versions_test.exs
@@ -19,7 +19,7 @@ defmodule GitHubActions.VersionsTest do
 
   test "from_config/0" do
     assert List.last(Versions.from_config()) ==
-             [otp: ["24.0/3"], elixir: ["1.11.4", "1.12.0/3", "1.13.0/3"]]
+             [otp: ["24.0/3"], elixir: ["1.11.4", "1.12.0/3", "1.13.0/4"]]
   end
 
   describe "get/2" do
@@ -106,7 +106,8 @@ defmodule GitHubActions.VersionsTest do
                "1.13.0",
                "1.13.1",
                "1.13.2",
-               "1.13.3"
+               "1.13.3",
+               "1.13.4"
              ]
     end
 
@@ -224,7 +225,8 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 13, patch: 0},
                %Version{major: 1, minor: 13, patch: 1},
                %Version{major: 1, minor: 13, patch: 2},
-               %Version{major: 1, minor: 13, patch: 3}
+               %Version{major: 1, minor: 13, patch: 3},
+               %Version{major: 1, minor: 13, patch: 4}
              ]
     end
 
@@ -257,7 +259,7 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 10, patch: 4},
                %Version{major: 1, minor: 11, patch: 4},
                %Version{major: 1, minor: 12, patch: 3},
-               %Version{major: 1, minor: 13, patch: 3}
+               %Version{major: 1, minor: 13, patch: 4}
              ]
     end
   end
@@ -373,7 +375,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 21, minor: 3},
-                 elixir: %Version{major: 1, minor: 13, patch: 3}
+                 elixir: %Version{major: 1, minor: 13, patch: 4}
                ],
                [
                  otp: %Version{major: 24, minor: 3},